requirements.txt
fastapi
uvicorn
python-telegram-bot
google-generativeai
playwright
nest_asyncio
python-dotenv
requests
